add async namespace, add bullet and enemy
