+++ Lava Vanguard/Assets/Scripts/Async.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ff3fafcfd8475964fb605b872fa9b8c1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
