+++ Lava Vanguard/Assets/Scripts/Async/GameManager/GameManager.cs
@@ -1,0 +1,24 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+namespace Async
+{
+    //Just for temp testing
+    public class GameManager : MonoBehaviour
+    {
+
+        public Transform draggingTransform;
+
+        public static GameManager Instance { get; private set; }
+        private void Awake()
+        {
+            Instance = this;
+        }
+        private void Start()
+        {
+            SequenceManager.Instance.Init();
+            InventoryManager.Instance.Init();
+        }
+
+    }
+}
