+++ Lava Vanguard/Assets/Scripts/Async/Inventory/InventoryManager.cs
@@ -1,0 +1,22 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace Async
+{
+    public class InventoryManager : MonoBehaviour
+    {
+        public static InventoryManager Instance { get; private set; }
+        public InventoryView inventoryView;
+
+        private void Awake()
+        {
+            Instance = this;
+        }
+        public void Init()
+        {
+            inventoryView.Init();
+        }
+
+    }
+}
