+++ Lava Vanguard/Assets/Scripts/Async/ScriptableObjects/CardConfig.cs
