+++ Lava Vanguard/Assets/Scripts/Async/ScriptableObjects/InventoryConfig.cs
