+++ Lava Vanguard/Assets/Scripts/Bullet.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d25340bcda90f184aaf2257622490f3d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
