+++ Lava Vanguard/Assets/Scripts/Bullet/BulletData.cs
@@ -1,0 +1,11 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class BulletData 
+{
+    //Level?
+    public string ID { get; set; }
+    public string Sprite { get; set; }
+
+}
