+++ Lava Vanguard/Assets/Scripts/Bullet/BulletData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7c20dccd31a33bc45b3e579201b03534
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
