+++ Lava Vanguard/Assets/Scripts/Bullet/BulletManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f02d77726166a2e4bbebe7eab54b7739
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
