+++ Lava Vanguard/Assets/Scripts/Bullet/BulletView.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f871ded237c249f4b983e36f64a98925
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
