+++ Lava Vanguard/Assets/Scripts/Enemy/EnemyData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7d42695131b8ca4419c59e0112afdf8d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
