+++ Lava Vanguard/Assets/Scripts/Enemy/EnemyManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 33e5236421af767419bb7b8817131074
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
