+++ Lava Vanguard/Assets/Scripts/Enemy/EnemyView.cs
@@ -1,0 +1,17 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class EnemyView : MonoBehaviour
+{
+    public EnemyData enemyData;
+    public SpriteRenderer spriteRenderer;
+    public void Init(EnemyData enemyData)
+    {
+
+    }
+    private void Update()
+    {
+        // Looming
+    }
+}
