+++ Lava Vanguard/Assets/Scripts/Enemy/EnemyView.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 148a2bb4f024aa847a119edcf4101d07
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
