--- Lava Vanguard/Assets/Scripts/Inventory/InventoryManager.cs
@@ -1,19 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class InventoryManager : MonoBehaviour
-{
-    public static InventoryManager Instance {  get; private set; }
-    public InventoryView inventoryView;
-   
-    private void Awake()
-    {
-        Instance = this; 
-    }
-    public void Init()
-    {
-        inventoryView.Init();
-    }
-
-}
